Merged PR 40704: Contact Update / Subscription Flow action Controller and Client methods
## What's being changed
We have added a new flow action to create, update and resubscribe contacts off the back of Shopware event triggers.

## Why it's being changed
Merchants would like the ability to add contacts to Dotdigital, optionally with data fields, within Shopware Flow Builder.

## How to review / test this change
1. Add Dotdigital Contact  action to a flow.
    - Update address book to a valid address book ID form your account within the component tempData object
    - Add valid data fields to the component temp data, update the other fields.
    - Toggle between different opt In and subscribe states
2. Add a break point `Dotdigital\Flow\Core\Content\Flow\Dispatching\Action\DotdigitalContactAction` @L85
3. Trigger your newly added action
4. Step through logic to ensure the handler does not fail.
5. Ensure the configurations you set are reflected within the EC platform.

 ## Notes
You will need to remove and add the action again if you update any of the temp data in the component before it will reflect on the handler.

Related work items: #189442, #189443