Merged PR 40842: Provide data via controller to the admin
## What's changed
In this PR we created API endpoints to fetch dataFields and addressBooks in the admin vue js components.
More specifically we added:
- 2 new controllers one for AddressBooks one for DataFIelds that uses DotdigitalClient to perform API requests to Dotdigital
- new javascript implementation of DotdigitalApiService to perform request in Shopware backend using the new controllers

## Why it's changed
We need to provide real dataFields and addressBooks in contact flow action modal.

## How to review
- Make sure you have valid dotdigital credentials in settings
- Go to flow builder and add a new contact flow action
- Open the browser logs and confirm that you get your addressBooks and dataFields

**Extra review steps**
Since we decided to refactor our code and to create an addition ContactData DataType we should run again the following steps
- https://dev.azure.com/dotdigital/ec/_git/dd-shopware-flowbuilder/pullrequest/40704

Related work items: #190678