Merged PR 44199: Compatibility fixes for Shopware 6.5
## What's being changed

This PR updates our code in preparation for the release of Shopware 6.5.

## Why it's being changed

Shopware 6.5 makes significant changes to the underlying code for processing Flow Builder actions, so we've had to adapt to these. We also needed to replace our usage of the old `@RouteScope` annotation, update some CSS styling and an icon name.

## How to review / test this change

- Review the UI for all 3 types of flow action
- Test variable usage in recipients, data fields and contact personalisation
- Refer back to the QA documents for this module for end-to-end testing

## Notes

- [Shopware 6.5 upgrade instructions](https://github.com/shopware/platform/blob/6.5.0.0/UPGRADE-6.5.md).

Related work items: #192958, #202892, #206952