Merged PR 44341: Remove MailAware-related exceptions and fix unit tests
## What's being changed

Any exceptions we used to throw relating to `MailAware` instances have been removed. Unit tests are also now passing.

## Why it's being changed

We used to throw exceptions if an event was not an instance of the MailAware class in our action classes - but since we no longer handle events (we get a `StorableFlow` instance) this class-checking no longer seems to have any purpose.

## How to review / test this change

- Run `vendor/bin/phpunit`

Related work items: #208886