Merged PR 46900: Insert consent field and validate it in user registration
## What's being changed

In this PR we added a new consent field in user registration that includes a validation of it's mobile phone number too.

## Why it's being changed

As part of the project london for shopware we need to allow users tho getting subscribed for sms marketing consent at user registration.

## How to review / test this change

- Register an account.
- see the new consent field.
- if checkbox is not checked then mobile field should not be visible and no validation is happening when submitting the registration button.
- If consent checkbox is checked then we should be able to see the mobile form film that includes the intlTelInput library and the dropdown with country flags. At this point you should carefully check the validation.
- You need to check the default country settings. Based on your storefront default country and countries list you should get the correct flags in the dropdown.

## Notes

This PR includes only UI changes

Related work items: #219699