Merged PR 45250: Fix PHP coding standard and JS linting errors
## What's being changed

We have run the fixer command on the plugin, we have added Eslint to plugin, as well as resolving all PHPStan issues.

## Why it's being changed

Coding standards help you develop less complex software and, therefore, reduce errors. Improvement of the bug fixing process: with coding standards, it becomes easier for developers to locate and correct bugs in the source code because it is written in a consistent manner

## How to review / test this change

1. Observe a pass on the Ecs tests
2. Observe a pass on eslinting
3. Observe a pass on PHP stan
4. Observe a pass on Unit tests
5. Do some Smoke tests of the functionality

Related work items: #211565