Merged PR 47323: Add Consent field in checkout
## What's being changed

A new consent checkbox added in the checkout page under the shipping address section.
A new partial created to contain consent field html code in order to be reusable in all places.
In this PR we use this in checkout and in user registration form.

## Why it's being changed

We are rolling out SMS consent capture at various customer touchpoints in Shopware.

## How to review / test this change

- Add a product to a cart and go to checkout page
- Confirm that a new consent field is now added under the shipping address section
- Ensure that the field is validated when submitting the order
- Confirm that this works for both customer and guest checkout

Related work items: #219716