Merged PR 47621: Fix missing value in extension configuration area
## What's being changed

We have set the data type to string within the AddressBook struct.

## Why it's being changed

The ECS runner sets all comparators to strict. The Shopware core entity resolver / xriteria object expects the value of ID to be type `string`, in the Repository reader the in_array was set to strict by ECS and in turn was not able to locate the set value due to the struct using an int and the criteria expecting a string.

## How to review / test this change

- Ensure selected list is visible in the shopware extensions page for dotdigital.
- Smoke test for Address Book ID

Related work items: #224048